Refactored and swizzle form.json portal URLs
